✨ feat(adaptive-learning): Initialize adaptive learning system
This commit introduces the foundational components and initialization logic for the
adaptive learning system, making it conditionally available based on the
application configuration. This system is crucial for enabling dynamic adjustment
of model weights and feedback processing.

**Changes made:**
-   **Enable** the **adaptive learning system** initialization in `cmd/serve.go`,
    creating repositories, scorers, learners, schedulers, and API handlers.
    The system is activated only when `cfg.AdaptiveLearning.Enabled` is true.
-   **Integrate** `slog` with the existing `zerolog` logging infrastructure
    within `cmd/serve.go` by introducing `slogFromZerolog` and
    `zerologSlogHandler`. This ensures that the learning scheduler, which
    uses `slog`, logs through the application's unified `zerolog` setup.
-   **Update** the `LearningFeedbackRepository.CountByAction` interface method
    in `internal/api/learning_handlers.go` to return `int64` instead of `int`.
    This change accommodates potentially larger counts of feedback actions.
-   **Adjust** the mock implementation of `CountByAction` in
    `internal/api/learning_handlers_test.go` to reflect the `int64` return type.
-   **Add** an explicit comment block for "Adaptive Learning Metrics" in
    `internal/metrics/metrics.go`, preparing the ground for future metric
    definitions related to the new system.

> **Note:** The adaptive learning system's core components, including
> `feedbackRecordRepo`, `weightHistoryRepo`, `scorer`, `weightLearner`, and
> `learningScheduler`, are now instantiated during application startup,
> allowing for dynamic learning capabilities.

**Technical details:**
-   Files modified: `cmd/serve.go`, `internal/api/learning_handlers.go`,
    `internal/api/learning_handlers_test.go`, `internal/metrics/metrics.go`.
-   `cmd/serve.go`:
    -   Added `log/slog` and `pharmabroker/internal/metrics` imports.
    -   Introduced `learningHandlers`, `learningScheduler` variables.
    -   Implemented conditional initialization for adaptive learning components.
    -   Added `slogFromZerolog` and `zerologSlogHandler` functions.
    -   Ensured `learningScheduler` is stopped during graceful shutdown.
-   `internal/api/learning_handlers.go`:
    -   Changed `CountByAction` signature from `(int, error)` to `(int64, error)`.
-   `internal/api/learning_handlers_test.go`:
    -   Modified `mockLearningFeedbackRepo.CountByAction` to return `int64`.
-   `internal/metrics/metrics.go`:
    -   Added `// ========== Adaptive Learning Metrics ==========` comment.

+++ internal/api/learning_handlers.go
@@ -20,7 +20,7 @@ type LearningHandlers struct {
 // LearningFeedbackRepository interface for learning feedback
 type LearningFeedbackRepository interface {
 	GetFeedbackStats(ctx context.Context, startDate, endDate time.Time) (*domain.FeedbackStats, error)
-	CountByAction(ctx context.Context, action domain.FeedbackAction, startDate, endDate time.Time) (int, error)
+	CountByAction(ctx context.Context, action domain.FeedbackAction, startDate, endDate time.Time) (int64, error)
 }
 
 // LearningWeightHistoryRepository interface for weight history
